backend/requirements.txt
# ───── Web framework + ASGI server ───────────────────────────
fastapi~=0.115
uvicorn[standard]~=0.35

# ───── Database driver ───────────────────────────────────────
motor~=3.4

# ───── Messaging / Kafka client ──────────────────────────────
kafka-python~=2.0

# ───── HTTP client (sync & async) ────────────────────────────
httpx~=0.27
requests~=2.32

# ───── Validation & helpers ──────────────────────────────────
pydantic~=2.7
python-dotenv~=1.0
typing-extensions>=4.0

# ──────  Test / CI / Lint ──────
pytest~=8.2
pytest-asyncio~=0.23
mongomock~=4.0
pytest-cov~=5.0
flake8~=7.0
black~=24.3
mypy~=1.10
